# Optimized range