--- a/brigand/sequences/range.hpp
+++ b/brigand/sequences/range.hpp
@@ -18,7 +18,7 @@ namespace brigand
     template<class T, class, class, T>
     struct range_cat;
 
-#ifndef BRIGAND_COMP_MSVC
+#ifdef BRIGAND_COMP_MSVC
     template<class T, T Start, T Int>
     struct int_plus
     {
@@ -29,10 +29,10 @@ namespace brigand
     template<class T, class... Ts, T... Ints, T Start>
     struct range_cat<T, list<Ts...>, list<std::integral_constant<T, Ints>...>, Start>
     {
-#ifndef BRIGAND_COMP_MSVC
-      using type = list<Ts..., std::integral_constant<T, Start + Ints>...>;
-#else
+#ifdef BRIGAND_COMP_MSVC
       using type = list<Ts..., typename int_plus<T, Start, Ints>::type...>;
+#else
+      using type = list<Ts..., std::integral_constant<T, Start + Ints>...>;
 #endif
     };
 
@@ -62,16 +62,22 @@ namespace brigand
     template<class T, class, class, T>
     struct reverse_range_cat;
 
+#ifdef BRIGAND_COMP_MSVC
     template<class T, T Start, T Int>
     struct int_minus
     {
       using type = std::integral_constant<T, Int - Start>;
     };
+#endif
 
     template<class T, class... Ts, T... Ints, T Start>
     struct reverse_range_cat<T, list<Ts...>, list<std::integral_constant<T, Ints>...>, Start>
     {
+#ifdef BRIGAND_COMP_MSVC
       using type = list<Ts..., typename int_minus<T, Start, Ints>::type...>;
+#else
+      using type = list<Ts..., std::integral_constant<T, Ints - Start>...>;
+#endif
     };
 
     template<class T, T Start, std::size_t N>