```cpp
using list1 = brigand::range<int, 0, SIZE>;
using list2 = brigand::range<long, 0, SIZE>;
using list3 = brigand::range<long long, 0, SIZE>;
using list4 = brigand::range<unsigned long long, 0, SIZE>;
```

Gcc:

Size = |      10      |      50      |     100      |     200      |     400
------ | ------------ | ------------ | ------------ | ------------ | ------------
before | 00.04s - 16M | 00.05s - 17M | 00.05s - 18M | 00.07s - 20M | 00.11s - 24M
after  | 00.03s - 16M | 00.04s - 17M | 00.04s - 17M | 00.05s - 18M | 00.06s - 19M
gain ~ |   25%  - 0% |   20% - 3% |   20% - 6% |   28% - 11% |   45% - 20%

Clang:

Size = |      10      |      50      |     100      |     200      |     400
------ | ------------ | ------------ | ------------ | ------------ | ------------
before | 00.06s - 36M | 00.06s - 36M | 00.07s - 37M | 00.09s - 38M | 00.11s - 41M
after  | 00.06s - 36M | 00.06s - 36M | 00.06s - 36M | 00.07s - 37M | 00.08s - 39M
gain ~ |    0% - 0% |    0% - 0% |   14% - 2% |   22% - 2% |   27% - 4%